Refactor bot

+++ src/handleData/index.js
@@ -1,3 +1,4 @@
+/* eslint-disable max-len */
 import * as cheerio from 'cheerio';
 import Axios from 'axios';
 import Track from '../../db/models/tracks';
@@ -24,22 +25,42 @@ export const getTrackInfo = async (trackNumber) => {
   return events;
 };
 
-export const addOrUpdateData = async ({ data, userId, trackingNumber }) => Track.updateOne(
-  { userId, trackingNumber },
-  { trackingNumber, userId, trackInfo: JSON.stringify(data) },
-  { upsert: true, setDefaultsOnInsert: true },
-);
+// export const addOrUpdateData = async ({
+//   data,
+//   userId,
+//   trackingNumber,
+//   ...rest
+// }) => {
+//   return Track.updateOne(
+//     { userId, trackingNumber },
+//     {
+//       trackingNumber,
+//       userId,
+//       trackInfo: JSON.stringify(data),
+//       ...rest,
+//     },
+//     { upsert: true, setDefaultsOnInsert: true },
+//   );
+// };
 
-export const deleteData = async ({ userId, trackingNumber }) => Track.deleteOne(
-  { userId, trackingNumber },
-);
+export const deleteData = async ({ userId, trackingNumber }) => Track.deleteOne({ userId, trackingNumber });
 
-export const saveData = async ({ userId, trackingNumber }) => {
-  const data = await getTrackInfo(trackingNumber);
-  try {
-    await addOrUpdateData({ data, userId, trackingNumber });
-    return { data, error: null };
-  } catch (error) {
-    return { data: null, error };
-  }
-};
+// export const saveData = async ({ userId, trackingNumber, parsedEvent = null, ...rest }) => {
+//   const data = await getTrackInfo(trackingNumber);
+
+//   const lastEvent = parsedEvent || data.length ? data.slice(-1)[0].event : null;
+
+//   try {
+//     await addOrUpdateData({
+//       data,
+//       userId,
+//       trackingNumber,
+//       lastUpdated: new Date(),
+//       lastEvent,
+//       ...rest,
+//     });
+//     return { data, error: null };
+//   } catch (error) {
+//     return { data: null, error };
+//   }
+// };
